Move form functions to controllers
