+++ controllers/content/create.post.js
@@ -5,8 +5,8 @@ const is = require('is_js')
 const path = require('path')
 const { DateTime } = require('luxon')
 
-const normalizeFormState = require('../../utilities/form-normalize-state')
-const normalizeFormErrors = require('../../utilities/form-normalize-errors')
+const normalizeFormState = require('./shared/form-normalize-state')
+const normalizeFormErrors = require('./shared/form-normalize-errors')
 
 const md = require('../../utilities/markdown-it')
 const config = require('../../config')
