+++ utilities/index.js
@@ -3,8 +3,6 @@ const combineMerge = require('./combine-merge')
 const constructOauthCallback = require('./construct-oauth-callback')
 const AppError = require('./app-error')
 const formGeo = require('./form-geo')
-const formNormalizeErrors = require('./form-normalize-errors')
-const formNormalizeState = require('./form-normalize-state')
 const markdownIt = require('./markdown-it')
 const monthNameToMonthNumber = require('./month-name-to-month-number')
 
@@ -14,8 +12,6 @@ module.exports = {
   constructOauthCallback,
   AppError,
   formGeo,
-  formNormalizeErrors,
-  formNormalizeState,
   markdownIt,
   monthNameToMonthNumber
 }
